refactored tests
