+++ exasol_integration_test_docker_environment/test/conftest.py
@@ -8,6 +8,7 @@ from exasol_integration_test_docker_environment.testing \
 from exasol_integration_test_docker_environment.testing \
    .exaslct_test_environment import SpawnedTestEnvironments
 
+# from fake import SpawnedTestEnvironments, ExaslctTestEnvironment, utils
 
 @pytest.fixture
 def itde_cli_test_isolation(request) -> Iterator[ExaslctTestEnvironment]:
@@ -22,19 +23,3 @@ def itde_cli_test_isolation(request) -> Iterator[ExaslctTestEnvironment]:
         yield environment
     finally:
         utils.close_environments(environment)
-
-
-@contextlib.contextmanager
-def database(itde_test_isolation: ExaslctTestEnvironment,
-             name: Optional[str] = None,
-             additional_parameters=Optional[List[str]] = None
-             ) -> Iterator[SpawnedTestEnvironments]:
-    name = name if name else itde_test_isolation.name
-    spawned = itde_test_isolation.spawn_docker_test_environments(
-        name=name,
-        additional_parameter=additional_parameters,
-    )
-    try:
-        yield spawned
-    finally:
-        utils.close_environments(spawned)
